pyproject.toml
[project]
name = "myproject"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi[standart]>=0.128.0",
    "flagembedding>=1.3.5",
    "hf-xet>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-deepseek>=1.0.1",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "mammoth>=1.11.0",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "pydantic>=2.12.5",
    "pymupdf4llm>=0.2.7",
    "python-magic-bin>=0.4.14",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
